src/requirements.txt

azureml==0.2.7
azureml_core==1.44.0
mlflow==2.0.1
mlflow_skinny==1.27.0
numpy==1.21.6
pandas==1.1.5
scikit_learn==1.1.3
xgboost==1.3.3
